Gui Layout Completed
The entire GUI layout is done, and working properly as i tested them. However, the add button on mainMenu frame for add a new reading need to be implemented.

for manage fiel frame, upload file need to be reworked so it wont add duplicated patients and save file button need to be implemented